requirements.txt
black==25.1.0
python-dotenv==1.1.0
python-telegram-bot==22.3
web3==7.13.0
safe-eth-py==7.9.0